// src/modules/enums/ui.enum.ts
export enum Page {
  HOME = 'Home',
  LANDING = 'Landing',
  ERROR = 'Error',
  LOGIN = 'Login',
  SIGNIN = 'SignIn',
  SIGNUP = 'Register',
  TOS = 'Terms Of Service',
  PRIVACY_POLICY = 'Privacy Policy',
}

export enum Path {
  HOME = '/',
  LANDING = '/landing',
  ERROR = '/error',
  TOS = '/tos',
  LOGIN = '/private/admin/page/login',
  SIGNIN = '/private/admin/page/signin',
  SIGNUP = '/private/admin/page/register',
  PRIVACY_POLICY = '/privacy',
}

export enum SchemeId {
  DARK = 'dark',
  LIGHT = 'light',
  SYSTEM = 'system',
}

export enum MixBlendMode {
  NORMAL = 'mix-blend-normal',
  MULTIPLY = 'mix-blend-multiply',
  SCREEN = 'mix-blend-screen',
  OVERLAY = 'mix-blend-overlay',
  SOFT_LIGHT = 'mix-blend-soft-light',
  HARD_LIGHT = 'mix-blend-hard-light',
  DIFFERENCE = 'mix-blend-difference',
}